% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{addInconsistentPositionAlignment}
\alias{addInconsistentPositionAlignment}
\title{Adds nucleotide alignment to right of plotted phylogenetic tree - nucleotides for inconsistent positions}
\usage{
addInconsistentPositionAlignment(tree, fastaFile, inconsistentPositions,
  cex, addSeparatorLines, actualPositions)
}
\arguments{
\item{tree}{An object of class "phylo" produced by \code{readAnnotatedTree()}}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{inconsistentPositions}{An integer array of the inconsistent positions identified by HomoplasyFinder. Produced by \code{runHomoplasyFinderInJava()}}

\item{cex}{A multiplier value to change the size of position labels}

\item{addSeparatorLines}{A boolean variable to determine whether white separator lines are plotted on alignment. Defaults to true}

\item{actualPositions}{An integer array of length=\code{length(inconsistentPositions)} that reports the actual positions on the genome of inconsistent positions. Useful when \code{homoplasyFinder} was run on concatenated sites. Defaults to \code{NULL} meaning it is ignored.}
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
