% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{addInternalNodeLabels}
\alias{addInternalNodeLabels}
\title{Adds labels to plotted tree indicating the internal nodes associated with each inconsistent position}
\usage{
addInternalNodeLabels(tree)
}
\arguments{
\item{tree}{An object of class "phylo" produced by \code{readAnnotatedTree()}}
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
