% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{addScaleBar}
\alias{addScaleBar}
\title{Adds a scale bar to plotted phylogenetic tree if requested}
\usage{
addScaleBar(addScale)
}
\arguments{
\item{addScale}{A flag to check whether scale wanted}
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
