% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{areSetsOfIsolatesTheSame}
\alias{areSetsOfIsolatesTheSame}
\title{Compare two vectors of isolate names}
\usage{
areSetsOfIsolatesTheSame(a, b)
}
\arguments{
\item{a}{An object of class "vector" containing character strings recognising isolates}

\item{b}{An object of class "vector" containing character strings recognising isolates}
}
\value{
Returns either TRUE or FALSE depending on whether the input vectors are identical or not
}
\description{
Function used by \code{assignAlellesToNodes()}
}
\keyword{internal}
