% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{assignAllelesAtPositions}
\alias{assignAllelesAtPositions}
\title{A single thread of assigning alleles to nodes in a phylogeny for selected positions}
\usage{
assignAllelesAtPositions(nodes, alleles, isolates, positions, verbose)
}
\arguments{
\item{nodes}{An object of class "list" recording the tips (isolates) associated with each node in a phylogeny - produced by \code{getNodes()}}

\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}

\item{isolates}{An object of class "vector" containing "character" strings identifying isolates}

\item{positions}{A vector of characters representing positions in a sequence alignment to be evaluated}

\item{verbose}{Print detailed progress information. Used only if run on a single thread. If run in parallel, it is ignored and set to FALSE.}
}
\value{
Returns an object of class "vector" with the IDs of alleles (Position:Nucleotide) not assigned to nodes on the phylogeny
}
\description{
Function used by \code{assignAllelesToNodes()} to assign alleles (identified by an ID (Position:Nucleotide)) to nodes in a phylogeny
}
\keyword{internal}
