% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{assignAllelesToNodes}
\alias{assignAllelesToNodes}
\title{Assign alleles to nodes in a phylogeny}
\usage{
assignAllelesToNodes(nodes, alleles, isolates, verbose,
  clusterOfThreads = NULL)
}
\arguments{
\item{nodes}{An object of class "list" recording the tips (isolates) associated with each node in a phylogeny - produced by \code{getNodes()}}

\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}

\item{isolates}{An object of class "vector" containing "character" strings identifying isolates}

\item{verbose}{Print detailed progress information}

\item{clusterOfThreads}{An object of class "cluster" to perform parallel computation. If NULL, the computation is performed on a single thread. Defaults to NULL}
}
\value{
Returns an object of class "vector" with the IDs of alleles (Position:Nucleotide) not assigned to nodes on the phylogeny
}
\description{
Function used by \code{homoplasyFinder()} to assign alleles (identified by an ID (Position:Nucleotide)) to nodes in a phylogeny
}
\keyword{internal}
