% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{checkForSpaces}
\alias{checkForSpaces}
\title{Check for spaces in file path - if found this function returns path surrounded by double quotes}
\usage{
checkForSpaces(filePath)
}
\arguments{
\item{filePath}{A path to a file}
}
\value{
Returns A character string of the file path surrounded by quotations
}
\description{
Function used by \code{runHomoplasyFinderJavaTool()}
}
\keyword{internal}
