% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getAllelePositions}
\alias{getAllelePositions}
\title{Get a vector of the unique positions across the alleles}
\usage{
getAllelePositions(alleles)
}
\arguments{
\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}
}
\value{
Returns a vector of the unique positions across the alleles
}
\description{
Function used by \code{assignAlellesToNodes()}
}
\keyword{internal}
