% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getAllelesAtPosition}
\alias{getAllelesAtPosition}
\title{Get a vector of the alleles found at a given position that haven't already been assigned to a node}
\usage{
getAllelesAtPosition(position, alleles)
}
\arguments{
\item{position}{A character string detailing the position of interest}

\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}
}
\value{
Returns a vector of unassigned alleles found at the position provided
}
\description{
Function used by \code{assignAlellesToNodes()}
}
\keyword{internal}
