% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getInternalNodeCoordinates}
\alias{getInternalNodeCoordinates}
\title{Get the X and Y coordinates of each internal node in the current plot.phylo plot}
\usage{
getInternalNodeCoordinates(nTips)
}
\arguments{
\item{nTips}{The number of tips in in the plotted tree. Calculate using (\code{length(tree$tip.label)})}
}
\value{
Returns an object of class "matrix" containing the X and Y coordinates of each internal node in the currently plotted phylogenetic tree
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
