% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getIsolatesWithNsAtPosition}
\alias{getIsolatesWithNsAtPosition}
\title{Get a vector of the isolates (associated with sequences) that have an N at the given allele's position}
\usage{
getIsolatesWithNsAtPosition(position, alleles)
}
\arguments{
\item{position}{A character string detailing the position of interest}

\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}
}
\value{
Returns a vector of the names of isolates that have an N at the position associated with the allele
}
\description{
Function used by \code{assignAlellesToNodes()}
}
\keyword{internal}
