% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getNodes}
\alias{getNodes}
\title{Record the tips associated with every node in a phylogeny}
\usage{
getNodes(tree)
}
\arguments{
\item{tree}{An object of class "phylo"}
}
\value{
Returns an object of class "list" detailing the isolates associated with every node in the input tree
}
\description{
Function used by \code{homoplasyFinder()}
}
\keyword{internal}
