% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getPositions}
\alias{getPositions}
\title{Get the positions from allele IDs}
\usage{
getPositions(alleles)
}
\arguments{
\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}
}
\value{
Returns a vector of the positions associated with the input allele information
}
\description{
Function used by \code{reportHomoplasiesIdentified()} to extract the positions of each allele id (Position:Nucleotide)
}
\keyword{internal}
