% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getTipCoordinates}
\alias{getTipCoordinates}
\title{Get the X and Y coordinates of each tip in the current plot.phylo plot}
\usage{
getTipCoordinates(tipLabels)
}
\arguments{
\item{tipLabels}{An object of class "vector" containing the labels assigned to each tip in the plotted tree. Get using (\code{tree$tip.label})}
}
\value{
Returns an object of class "list" containing the X and Y coordinates of each tip in the currently plotted phylogenetic tree
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
