% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{getTipSequences}
\alias{getTipSequences}
\title{Reads in FASTA file and records which sequences is associated with each tip label}
\usage{
getTipSequences(fastaFile)
}
\arguments{
\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}
}
\value{
Returns an object of class "list" containing the nucleotide sequence associated with each tip in the plotted phylogenetic tree
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
