% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{homoplasyFinder}
\alias{homoplasyFinder}
\title{HomoplasyFinder - Identify homoplasies on phylogeny using its sequence alignment}
\usage{
homoplasyFinder(tree, sequencesDNABin, verbose = TRUE,
  nThreads = parallel::detectCores())
}
\arguments{
\item{tree}{An object of class "phylo"}

\item{sequencesDNABin}{An object containing FASTA formatted nucleotide sequences of class "DNAbin"}

\item{verbose}{Print detailed progress information. Defaults to TRUE}

\item{nThreads}{The number of parallel processes to use to perform the computation. Defaults to all available cores on the machine}
}
\value{
Returns a data.frame detailing the Positions, Alleles and Isolates associated with homoplasies identified
}
\description{
Function to run HomoplasyFinder - R version. Note considerably slower than java version
}
\examples{
# Load the tree and sequence data
data("tree")
data("sequences")

# Run homoplasy finder
results <- homoplasyFinder(tree, sequences)

# Plot the results
plotTreeAndHomoplasySites(tree, results)

}
\keyword{HomoplasyFinder}
\keyword{R}
\keyword{tool}
