% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{mergeListsByName}
\alias{mergeListsByName}
\title{Merge a list of named lists into one list, keeping keys and all elements of keys from all lists in l}
\usage{
mergeListsByName(lists)
}
\arguments{
\item{allelesFromEachThread}{list of named lists}
}
\value{
Returns a merged list with all keys.
}
\description{
Merge a list of named lists into one list, keeping keys and all elements of keys from all lists in l
}
\keyword{internal}
