% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{plotAnnotatedTree}
\alias{plotAnnotatedTree}
\title{Plotting HomoplasyFinder results}
\usage{
plotAnnotatedTree(
  tree,
  inconsistentPositions,
  fastaFile,
  addScale = FALSE,
  alignmentCex = 2,
  addNodeLabels = TRUE,
  nodeLabelCex = 1,
  alignmentPositionCex = 0.5,
  addSeparatorLines = TRUE,
  actualPositions = NULL
)
}
\arguments{
\item{tree}{An object of class "phylo" produced by \code{readAnnotatedTree()}}

\item{inconsistentPositions}{An integer array of the inconsistent positions identified by HomoplasyFinder. Produced by \code{runHomoplasyFinderInJava()}}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{addScale}{A boolean value to determine whether a scale is added to plot. Defaults to FALSE}

\item{alignmentCex}{A multiplier value to change the size of the space alotted to plotting the homoplasy alignment. Defaults to 2}

\item{addNodeLabels}{A multiplier value to change the size node labels. Defaults to 1}

\item{nodeLabelCex}{A multiplier value to change the size of the internal node labels added to phylogeny. Defaults to 1}

\item{alignmentPositionCex}{A multiplier value to change the size of the position labels added above alignment. Defaults to 0.5}

\item{addSeparatorLines}{A boolean variable to determine whether white separator lines are plotted on alignment. Defaults to true}

\item{actualPositions}{An integer array of length=\code{length(inconsistentPositions)} that reports the actual positions on the genome of inconsistent positions. Useful when \code{homoplasyFinder} was run on concatenated sites. Defaults to \code{NULL} meaning it is ignored.}
}
\description{
This function plots an annotated phylogeny with the inconsistent sites identified by HomoplasyFinder. Alignment colours: A=red, C=blue, G=cyan, and T=orange. NOTE: Works best with addTextLabels() package installed.
}
\examples{
# Find the FASTA and tree files attached to package
fastaFile <- system.file("extdata", "example.fasta", package = "homoplasyFinder")
treeFile <- system.file("extdata", "example.tree", package = "homoplasyFinder")

# Get the current working directory
workingDirectory <- paste0(getwd(), "/")

# Run the HomoplasyFinder java code
inconsistentPositions <- runHomoplasyFinderInJava(treeFile, fastaFile, workingDirectory)

# Get the current date
date <- format(Sys.Date(), "\%d-\%m-\%y")
 
# Read in the output table
resultsFile <- paste0(workingDirectory, "consistencyIndexReport_", date, ".txt")
results <- read.table(resultsFile, header=TRUE, sep="\\t", stringsAsFactors=FALSE)

# Read in the annotated tree
tree <- readAnnotatedTree(workingDirectory)

# Plot the annotated tree
plotAnnotatedTree(tree, inconsistentPositions, fastaFile)
}
\keyword{annotated}
\keyword{plot}
\keyword{tree}
