% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{plotTreeAndHomoplasySites}
\alias{plotTreeAndHomoplasySites}
\title{Plotting HomoplasyFinder results}
\usage{
plotTreeAndHomoplasySites(tree, results, scale = TRUE)
}
\arguments{
\item{tree}{An object of class "phylo"}

\item{results}{An object of class "data.frame" \code{homoplasyFinder()} or \code{runHomoplasyFinderJavaTool()}#}

\item{scale}{A boolean value to determine whether a scale is added to plot. Defaults to TRUE}
}
\description{
This function plots a phylogeny alongside the sites where homoplasies were identified
}
\examples{
# Load the tree and sequence data
data("tree")
data("sequences")

# Run homoplasy finder
results <- homoplasyFinder(tree, sequences)

# Plot the results
plotTreeAndHomoplasySites(tree, results)
}
\keyword{plot}
