% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{recordAllelesForSequences}
\alias{recordAllelesForSequences}
\title{Record each allele (Position:Nucleotide) found in the subset of sequences}
\usage{
recordAllelesForSequences(sequencesTable, verbose)
}
\arguments{
\item{verbose}{Print detailed progress information. Used only if run on a single thread. If run in parallel, it is ignored and set to FALSE.}

\item{sequences}{An object of class "data.frame" containing FASTA formatted nucleotide sequences}
}
\value{
Returns an object of class "list" detailing the isolates associated with every allele present in the sequences
}
\description{
Function used by \code{recordAllelesInPopulation()}
}
\keyword{internal}
