% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{recordAllelesInPopulation}
\alias{recordAllelesInPopulation}
\title{Record each allele (Position:Nucleotide) found in the set of sequences}
\usage{
recordAllelesInPopulation(sequences, verbose, clusterOfThreads = NULL)
}
\arguments{
\item{sequences}{An object of class "alignment" containing FASTA formatted nucleotide sequences}

\item{verbose}{Print detailed progress information.}

\item{clusterOfThreads}{An object of class "cluster" to perform parallel computation. If NULL, the computation is performed on a single thread. Defaults to NULL}
}
\value{
Returns an object of class "list" detailing the isolates associated with every allele present in the sequences
}
\description{
Function used by \code{homoplasyFinder()}
}
\keyword{internal}
