% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{removeConstantSites}
\alias{removeConstantSites}
\title{Remove constant sites}
\usage{
removeConstantSites(alleles, nSites, verbose = TRUE)
}
\arguments{
\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}

\item{nSites}{The number of positions in the nucleotide sequence alignment}
}
\description{
Function used by \code{homoplasyFinder()} to remove constant sites (only one nucleotide observed across sequences)
}
\keyword{internal}
