% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{reportHomoplasiesIdentified}
\alias{reportHomoplasiesIdentified}
\title{Summarise the results from homoplasyFinder() into results table}
\usage{
reportHomoplasiesIdentified(notAssigned, alleles)
}
\arguments{
\item{notAssigned}{An object of class "vector" containing the allele IDs (Position:Nucleotide) produced by \code{assignAlellesToNodes()}}

\item{alleles}{An object of class "list" recording the isolates associated with each allele produced by \code{recordAllelesInPopulation()}}
}
\value{
Returns a data.frame detailing the Positions, Alleles and Isolates associated with homoplasies identified
}
\description{
Function used by \code{homoplasyFinder()} to store the details of the homoplasy sites identified in dataframe
}
\keyword{internal}
