% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{runHomoplasyFinderInJava}
\alias{runHomoplasyFinderInJava}
\title{Run HomoplasyFinder using Java code}
\usage{
runHomoplasyFinderInJava(treeFile, fastaFile, path, createFasta = TRUE,
  createReport = TRUE, createAnnotatedNewickTree = TRUE,
  includeConsistentSitesInReport = FALSE, verbose = TRUE)
}
\arguments{
\item{treeFile}{The full path to the Newick formatted phylogenetic tree file}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{path}{The full path to the directory where the output files will be created}

\item{createFasta}{Flag to tell HomoplasyFinder whether to create FASTA file without any inconsistent positions. Defaults to TRUE}

\item{createReport}{Flag to tell HomoplasyFinder whether to create report file detailing the inconsistent sites identified. Defaults to TRUE}

\item{createAnnotatedNewickTree}{Flag to tell HomoplasyFinder whether to create annotated Newick formatted phylogenetic tree file. Defaults to TRUE}

\item{includeConsistentSitesInReport}{Flag to tell HomoplasyFinder whether to include information about the consistent sites in the report. Defaults to TRUE}

\item{verbose}{Flag to parse to HomoplasyFinder to request detailed information. Defaults to true}
}
\description{
This function runs HomoplasyFinder (coded in Java) to identify positions that are potentially homoplasious
}
\keyword{homoplasyFinder}
\keyword{java}
