% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{runHomoplasyFinderJarTool}
\alias{runHomoplasyFinderJarTool}
\title{Run HomoplasyFinder using Java jar file. Use if you can't get rJava to install properly :-)}
\usage{
runHomoplasyFinderJarTool(treeFile, fastaFile, createFasta = TRUE,
  createAnnotatedNewickTree = TRUE,
  includeConsistentSitesInReport = FALSE, verbose = TRUE)
}
\arguments{
\item{treeFile}{The full path to the Newick formatted phylogenetic tree file}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{createFasta}{Flag to tell HomoplasyFinder whether to create FASTA file without any inconsistent positions. Defaults to TRUE}

\item{createAnnotatedNewickTree}{Flag to tell HomoplasyFinder whether to create annotated Newick formatted phylogenetic tree file. Defaults to TRUE}

\item{includeConsistentSitesInReport}{Flag to tell HomoplasyFinder whether to include information about the consistent sites in the report. Defaults to TRUE}

\item{verbose}{Flag to parse to HomoplasyFinder to request detailed information. Defaults to true}
}
\description{
This function runs HomoplasyFinder as a command line jar tool to identify positions that are potentially homoplasious. Output(s) will appear in current working directory.
}
\keyword{homoplasyFinder}
\keyword{java}
