% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{runHomoplasyFinderJavaTool}
\alias{runHomoplasyFinderJavaTool}
\title{Run HomoplasyFinder.jar file}
\usage{
runHomoplasyFinderJavaTool(jarFile, fastaFile, treeFile, verbose = TRUE)
}
\arguments{
\item{jarFile}{A character string specifying the full path to HomoplasyFinder jar file}

\item{fastaFile}{A character string specifying the full path to FASTA sequences file}

\item{treeFile}{A character string specifying the full path to HomoplasyFinder jar file}

\item{verbose}{Flag to pass to HomoplasyFinder.jar to request detailed progress information. Defaults to TRUE}
}
\value{
Returns a data.frame detailing the Positions, Alleles and Isolates associated with homoplasies identified
}
\description{
Function to run HomoplasyFinder.jar file (java version of HomoplasyFinder)  within R - considerably faster than R package
}
\examples{
# Run homoplasyFinder
results <- runHomoplasyFinderJavaTool("fullPathToJarFile", "fullPathToFastaFile", "fullPathToTreeFile")

# Plot the results
plotTreeAndHomoplasySites(read.tree("fullPathToTreeFile"), results)
}
\keyword{HomoplasyFinder}
\keyword{jar}
\keyword{java}
\keyword{tool}
