% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\docType{data}
\name{sequences}
\alias{sequences}
\title{An object of class "phylo" detailing a phylogenetic tree}
\description{
An object of class "phylo" detailing a phylogenetic tree
}
\author{
Joseph Crispell \email{crispelljoseph@gmail.com}
}
\keyword{data}
