% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{splitVectorIntoParts}
\alias{splitVectorIntoParts}
\title{Split a vector into elements of a list in order to send each to a separate thread}
\usage{
splitVectorIntoParts(vector, nParts)
}
\arguments{
\item{vector}{An object of class "vector" to split into multiple parts}

\item{nParts}{An integer specifying the number parts the vector is to be split into}
}
\value{
Returns an object of class "list" of subsets of the input "vector"
}
\description{
Function used by \code{assignAlellesToNodes()}
}
\keyword{internal}
