% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{startCluster}
\alias{startCluster}
\title{Register cluster to start parallel computation}
\usage{
startCluster(nThreads)
}
\arguments{
\item{nThreads}{The number of parallel processes to use to perform the computation. Defaults to all available cores on the machine. If negative then that absolute number of threads will be left free}
}
\value{
Returns an object of class "cluster" to start parallel computation or NULL if computation will be performed on one thread
}
\description{
Function used by \code{homoplasyFinder()}
}
\keyword{internal}
