% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\docType{data}
\name{tree}
\alias{tree}
\title{An object of class "DNAbin" containing nucleotide sequence alignment}
\description{
An object of class "DNAbin" containing nucleotide sequence alignment
}
\author{
Joseph Crispell \email{crispelljoseph@gmail.com}
}
\keyword{data}
