% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{addInconsistentPositionAlignment}
\alias{addInconsistentPositionAlignment}
\title{Adds nucleotide alignment to right of plotted phylogenetic tree - nucleotides for inconsistent positions}
\usage{
addInconsistentPositionAlignment(tree, fastaFile, inconsistentPositions,
  cex, addSeparatorLines)
}
\arguments{
\item{tree}{An object of class "phylo" produced by \code{readAnnotatedTree()}}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{inconsistentPositions}{An integer array of the inconsistent positions identified by HomoplasyFinder. Produced by \code{runHomoplasyFinderInJava()}}

\item{cex}{A multiplier value to change the size of position labels}
}
\description{
Function used by \code{plotAnnotatedTree()}
}
\keyword{internal}
