% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{plotAnnotatedTree}
\alias{plotAnnotatedTree}
\title{Plotting HomoplasyFinder results}
\usage{
plotAnnotatedTree(tree, inconsistentPositions, fastaFile,
  addScale = TRUE)
}
\arguments{
\item{tree}{An object of class "phylo" produced by \code{readAnnotatedTree()}}

\item{inconsistentPositions}{An integer array of the inconsistent positions identified by HomoplasyFinder. Produced by \code{runHomoplasyFinderInJava()}}

\item{fastaFile}{The full path to the FASTA formatted nucleotide sequence alignment}

\item{scale}{A boolean value to determine whether a scale is added to plot. Defaults to TRUE}
}
\description{
This function plots an annotated phylogeny with the inconsistent sites identified by HomoplasyFinder
}
\keyword{annotated}
\keyword{plot}
\keyword{tree}
