% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{runHomoplasyFinderInJava}
\alias{runHomoplasyFinderInJava}
\title{Run HomoplasyFinder using Java code}
\usage{
runHomoplasyFinderInJava(
  treeFile,
  fastaFile = "Not provided",
  presenceAbsenceFile = "Not provided",
  path,
  createFasta = TRUE,
  createReport = TRUE,
  createAnnotatedNewickTree = TRUE,
  includeConsistentSitesInReport = FALSE,
  verbose = TRUE,
  multithread = FALSE
)
}
\arguments{
\item{treeFile}{The full path to the Newick formatted phylogenetic tree file}

\item{fastaFile}{The full path to a FASTA formatted nucleotide sequence alignment. Defaults to "Not provided"}

\item{presenceAbsenceFile}{The full path to a CSV table file reporting the presence/absence of INDELs. Defaults to "Not provided"}

\item{path}{The full path to the directory where the output files will be created}

\item{createFasta}{Flag to tell HomoplasyFinder whether to create FASTA file without any inconsistent positions. Defaults to TRUE}

\item{createReport}{Flag to tell HomoplasyFinder whether to create report file detailing the inconsistent sites identified. Defaults to TRUE}

\item{createAnnotatedNewickTree}{Flag to tell HomoplasyFinder whether to create annotated Newick formatted phylogenetic tree file. Defaults to TRUE}

\item{includeConsistentSitesInReport}{Flag to tell HomoplasyFinder whether to include information about the consistent sites in the report. Defaults to TRUE}

\item{verbose}{Flag to parse to HomoplasyFinder to request detailed information. Defaults to true}

\item{multithread}{Flag to tell HomoplasyFinder to use multiple threads. Defaults to false}
}
\description{
This function runs HomoplasyFinder (coded in Java) to identify positions that are potentially homoplasious
}
\examples{
# Find the FASTA and tree files attached to package
fastaFile <- system.file("extdata", "example.fasta", package = "homoplasyFinder")
treeFile <- system.file("extdata", "example.tree", package = "homoplasyFinder")

# Get the current working directory
workingDirectory <- paste0(getwd(), "/")

# Run the HomoplasyFinder java code
inconsistentPositions <- runHomoplasyFinderInJava(treeFile, fastaFile, path=workingDirectory)

# Get the current date
date <- format(Sys.Date(), "\%d-\%m-\%y")
 
# Read in the output table
resultsFile <- paste0(workingDirectory, "consistencyIndexReport_", date, ".txt")
results <- read.table(resultsFile, header=TRUE, sep="\t", stringsAsFactors=FALSE)

# Read in the annotated tree
tree <- readAnnotatedTree(workingDirectory)

# Plot the annotated tree
plotAnnotatedTree(tree, inconsistentPositions, fastaFile)
}
\keyword{homoplasyFinder}
\keyword{java}
