% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasyFinder.R
\name{runHomoplasyFinderJarTool}
\alias{runHomoplasyFinderJarTool}
\title{Run HomoplasyFinder using Java jar file. Use if you can't get rJava to install properly :-)}
\usage{
runHomoplasyFinderJarTool(
  treeFile,
  fastaFile = NULL,
  presenceAbsenceFile = NULL,
  createFasta = TRUE,
  createAnnotatedNewickTree = TRUE,
  includeConsistentSitesInReport = FALSE,
  verbose = TRUE,
  multithread = FALSE
)
}
\arguments{
\item{treeFile}{The full path to the Newick formatted phylogenetic tree file}

\item{fastaFile}{The full path to a FASTA formatted nucleotide sequence alignment. Defaults to NULL}

\item{presenceAbsenceFile}{The full path to a CSV table file reporting the presence/absence of INDELs. Defaults to NULL}

\item{createFasta}{Flag to tell HomoplasyFinder whether to create FASTA file without any inconsistent positions. Defaults to TRUE}

\item{createAnnotatedNewickTree}{Flag to tell HomoplasyFinder whether to create annotated Newick formatted phylogenetic tree file. Defaults to TRUE}

\item{includeConsistentSitesInReport}{Flag to tell HomoplasyFinder whether to include information about the consistent sites in the report. Defaults to TRUE}

\item{verbose}{Flag to tell HomoplasyFinder to request detailed information. Defaults to true}

\item{multithread}{Flag to tell HomoplasyFinder to use multiple threads. Defaults to false}
}
\description{
This function runs HomoplasyFinder as a command line jar tool to identify positions that are potentially homoplasious. Output(s) will appear in current working directory.
}
\examples{
# Find the FASTA and tree files attached to package
fastaFile <- system.file("extdata", "example.fasta", package = "homoplasyFinder")
treeFile <- system.file("extdata", "example.tree", package = "homoplasyFinder")

# Run the HomoplasyFinder jar tool
runHomoplasyFinderJarTool(treeFile, fastaFile)

# Get the current date
date <- format(Sys.Date(), "\%d-\%m-\%y")

# Get the current working directory
workingDirectory <- paste0(getwd(), "/")

# Read in the output table
resultsFile <- paste0(workingDirectory, "consistencyIndexReport_", date, ".txt")
results <- read.table(resultsFile, header=TRUE, sep="\t", stringsAsFactors=FALSE)

# Note the inconsistent positions
inconsistentPositions <- results[results$ConsistencyIndex < 1, "Position"]

# Read in the annotated tree
tree <- readAnnotatedTree(workingDirectory)

# Plot the annotated tree
plotAnnotatedTree(tree, inconsistentPositions, fastaFile)
}
\keyword{homoplasyFinder}
\keyword{java}
